refactor:: move renderDefinitions to emitTypeSpec
